// WebApplication1/client/src/objects/ProfileInput.tsx
export type ProfileInput = {
    knowledge: string;
    interests: string;
    searchStatus: boolean;
    mentorSearchStatus: boolean;
    mentorStatus: boolean;
    description: string;
};

//фиксируем инфу с сервера
// в каком виде отправляем инфу на сервер